Change `Particle::id` back to a `usize`
A matrix graph representation is probabaly easier than a map
representation. Thus, the `Particle`s will need to be indexed.

+++ src/particle.rs
@@ -3,7 +3,7 @@
 use std::collections::HashMap;
 use std::fmt::Debug;
 use std::hash::Hash;
-use std::sync::atomic::{AtomicU32, Ordering};
+use std::sync::atomic::{AtomicUsize, Ordering};
 
 use crate::vector3d::Vector3d;
 
@@ -11,13 +11,13 @@ use crate::vector3d::Vector3d;
 /// Counter for the [`id`] property of the [`Particle`] class.
 ///
 /// [`id`]: Particle::id
-static PARTICLE_COUNTER: AtomicU32 = AtomicU32::new(0);
+static PARTICLE_COUNTER: AtomicUsize = AtomicUsize::new(0);
 
 
 /// A single particle in a longitudinal wave,
 /// each connected to other particles by linear springs.
 pub struct Particle {
-    id: u32,
+    id: usize,
     /// The mass of this particle in kilograms (kg).
     pub mass: f64,
     /// The position of this particle as a 3D vector in meters (m).
